requirements.txt
Django
djangorestframework
drf-nested-routers
redis
django-redis
psycopg2-binary
python-dotenv
aiohttp
tqdm